IAtaplink/requirements.txt
crewai[tools]>=0.80.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
toon-format>=0.9.0b1
